Basic Hybrid Music Implementation
A basic, still incomplete implementation of the Hybrid Music setting, which makes the game switch back to the default Wynncraft music when there is no Wynntils music in the current area.

Although the code is messy and almost certainly inefficient, before I started working on this I had spent less than an hour programming in java in my life. Because of this I am surprised that I even got it to work at all.

Note: The transitions are terrible and need to be fixed, but it should work.